Commit para adicionar alterações 26/12/2018

+++ lib/my_app_web/controllers/user_controller.ex
@@ -12,11 +12,16 @@ defmodule MyAppWeb.UserController do
   end
 
   def create(conn, %{"user" => user_params}) do
-    with {:ok, %User{} = user} <- Auth.create_user(user_params) do
-      conn
-      |> put_status(:created)
-      |> put_resp_header("location", Routes.user_path(conn, :show, user))
-      |> render("show.json", user: user)
+    case Auth.create_user(user_params) do
+      {:ok, %User{} = user} ->
+        conn
+        |> put_status(:created)
+        |> put_resp_header("location", Routes.user_path(conn, :show, user))
+        |> render("show.json", user: user)
+      {:error, _user} ->
+        conn
+        |> put_status(:unauthorized)
+        |> render(MyAppWeb.ErrorView, "401.json", message: "Unable to create user! Try again with other options!")
     end
   end
 
@@ -51,9 +56,24 @@ defmodule MyAppWeb.UserController do
 
       {:error, message} ->
         conn
-        |> delete_session(:current_user_id)
+        # |> delete_session(:current_user_id)
         |> put_status(:unauthorized)
         |> render(MyAppWeb.ErrorView, "401.json", message: message)
+    end
+  end
+
+  def logout(conn, _params) do
+    current_user_id = get_session(conn, :current_user_id)
+
+    if current_user_id do
+      conn
+      |> delete_session(:current_user_id)
+      |> render(MyAppWeb.UserView, "logout.json")
+    else
+      conn
+      |> put_status(:unauthorized)
+      |> render(MyAppWeb.ErrorView, "401.json", message: "Unauthenticated user")
+      |> halt()
     end
   end
 end
